refactor: move EVENT_SORT_OPTIONS and EventsOrderBySelect to components
Also move related hook useSetSortQueryParamToOptionValue to components.

refs TH-1294
