+++ apps/events-helsinki/src/domain/event/queryUtils.ts
@@ -7,16 +7,14 @@ import {
   useEventListQuery,
   getEventIdFromUrl,
   SIMILAR_EVENTS_AMOUNT,
+  EVENT_SORT_OPTIONS,
 } from 'events-helsinki-components';
 import { useTranslation } from 'next-i18next';
 import React from 'react';
 import { toast } from 'react-toastify';
 
 import AppConfig from '../app/AppConfig';
-import {
-  EVENT_SEARCH_FILTERS,
-  EVENT_SORT_OPTIONS,
-} from '../search/eventSearch/constants';
+import { EVENT_SEARCH_FILTERS } from '../search/eventSearch/constants';
 import {
   getEventCategories,
   getEventSearchVariables,
