+++ apps/events-helsinki/src/domain/search/eventSearch/__tests__/utils.test.ts
@@ -1,10 +1,9 @@
 import type { AppLanguage } from 'events-helsinki-components';
-import { DATE_TYPES } from 'events-helsinki-components';
+import { DATE_TYPES, EVENT_SORT_OPTIONS } from 'events-helsinki-components';
 import { advanceTo, clear } from 'jest-date-mock';
 
 import {
   EVENT_DEFAULT_SEARCH_FILTERS,
-  EVENT_SORT_OPTIONS,
   MAPPED_EVENT_CATEGORIES,
 } from '../constants';
 import { getEventSearchVariables, getNextPage, getSearchQuery } from '../utils';
