+++ packages/components/src/hooks/index.ts
@@ -14,6 +14,7 @@ export { default as useHomeTranslation } from './useHomeTranslation';
 export { default as useFooterTranslation } from './useFooterTranslation';
 export { default as useNotFoundTranslation } from './useNotFoundTranslation';
 export { default as useSearchTranslation } from './useSearchTranslation';
+export { default as useSetSortQueryParamToOptionValue } from './useSetSortQueryParamToOptionValue';
 export { default as useEventTranslation } from './useEventTranslation';
 export { default as useVenueTranslation } from './useVenueTranslation';
 export { default as useAppEventsTranslation } from './useAppEventsTranslation';
